agent/requirements.txt
openai>=1.50.0
anthropic>=0.40.0
requests>=2.31.0
opentelemetry-api>=1.40.0
opentelemetry-sdk>=1.40.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
httpx>=0.28.0
cryptography>=43.0.0
pyyaml>=6.0
pyjwt>=2.8.0
